fix: eliminar usuario (no solucionado)

+++ Backend_Up2Play/backend/src/main/java/com/Up2Play/backend/Service/NotificacionService.java
@@ -360,8 +360,11 @@ public class NotificacionService {
         UsuarioNotificacion usuarioNotificacion = usuarioNotificacionRepository.findByUsuarioAndNotificacion(
                 usuario,
                 notificacion);
-
-        usuarioNotificacionRepository.deleteById(usuarioNotificacion.getId());
+        
+        if (usuarioNotificacion != null) {
+            usuarioNotificacionRepository.deleteById(usuarioNotificacion.getId());
+            }
+        
 
         return true;
